Revert "Extract headers from requests (#4)" (#5)
This reverts commit 7d7ba27907d785c24c5ecf640911a78653585fff.
