+++ package.json
@@ -1,20 +1,16 @@
 {
-  "name": "@gallant/bulk-sms",
+  "name": "Gallant_Bulk_SMS",
   "version": "0.0.1",
   "main": "app.js",
   "dependencies": {
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "mysql2": "^3.10.2",
-    "sequelize": "^6.37.3",
-    "ua-parser-js": "^1.0.38"
+    "sequelize": "^6.37.3"
   },
   "scripts": {
     "start": "node app.js",
-    "devStart": "nodemon app.js",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "test": "jest --coverage --passWithNoTests"
+    "devStart": "nodemon app.js"
   },
   "type": "module",
   "keywords": [
@@ -26,11 +22,6 @@
   "license": "ISC",
   "description": "",
   "devDependencies": {
-    "@eslint/js": "^9.6.0",
-    "eslint": "^9.6.0",
-    "eslint-plugin-jest": "^28.6.0",
-    "globals": "^15.8.0",
-    "jest": "^29.7.0",
     "nodemon": "^3.1.4"
   }
 }
